refactor: change step number
